Crate npm_package_json

serde mappings for npm's package.json file format.

This does only implement the fields defined in the official npm documentation. It is common enough that packages define custom entries that are required by various tooling.

Structs

Bug
A bug contacting form.

Package
The top-level package.json structure.

Person
A person.

Repository
A repository.

Enums

Error
The errors that this library can return.

ManReference
A reference to a man page.

PersonReference
A reference to a person.

RepositoryReference
A repository reference.

Type Definitions

BinSet
An ordered map for bin entries.

DepsSet
An ordered map for dependencies entries.

EnginesSet
An ordered map for engines entries.

Result
The result type of this crate.

ScriptsSet
An ordered map for scripts entries.